Add config.py to map the env vars
