+++ app/main.py
@@ -1,23 +1,9 @@
-from functools import lru_cache
 from fastapi import FastAPI
-from pydantic_settings import BaseSettings, SettingsConfigDict
-
 from .routers import metrics
-
-
-class Settings(BaseSettings):
-    stac_url: str = "http://localhost:8080"
-    env: str = "dev"
-
-    model_config = SettingsConfigDict(env_file=".env")
-
-
-@lru_cache
-def get_settings():
-    return Settings()
-
+from .config import get_settings
 
 settings = get_settings()
+
 app = FastAPI(
     title="BioTableroSearch",
     description="Get metrics by predefined or custom (polygon) areas.",
@@ -30,7 +16,6 @@ app = FastAPI(
     },
     docs_url=None if settings.env.lower() == "prod" else "/docs",
 )
-
 
 app.include_router(metrics.router)
 # TODO: disable swagger on production: https://fastapi.tiangolo.com/tutorial/metadata/#docs-urls
