+++ requirements.txt
@@ -7,4 +7,4 @@ rasterio>=1.3.10,<1.4.0
 numpy>=1.26.4,<1.27.0
 matplotlib>=3.9.0,<3.10.0
 shapely>=2.0.4,<2.1.0
-Pillow>=10.3.0,<10.4.0
+python-dotenv>=1.0.1,<1.1.0
